moved redundant code from player and host components to game
