+++ src/components/game/Host.jsx
@@ -3,40 +3,22 @@ import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import {
   setCurrentTrack,
-  setGame,
-  setPlaying
 } from '../../actions/SpotifyActions';
 import styles from './Host.css';
 
-export const Host = () => {
+const Host = () => {
   const {
     socket,
-    currentTrack,
     tracks,
     game,
     playing,
+    currentTrack
   } = useSelector(state => state);
   const [guess, setGuess] = useState(''); 
   const [player, setPlayer] = useState({}); 
   const dispatch = useDispatch();
+  
   useEffect(() => {
-    if(!socket) return;
-    socket.on('CHANGE_TRACK', track => {
-      dispatch(setCurrentTrack(track));
-    });
-    socket.on('PAUSE', () => {
-      const audioEl = document.getElementById('player');
-      audioEl.pause();
-      dispatch(setPlaying(false));
-    });
-    socket.on('PLAY', () => {
-      const audioEl = document.getElementById('player');
-      audioEl.play();
-      dispatch(setPlaying(true));
-    }); 
-    socket.on('GAME_INFO', ({ game }) => {
-      dispatch(setGame(game));
-    });
     socket.on('GUESS', ({ userGuess, user, playerId }) => {
       setGuess(userGuess);
       setPlayer({ name:user, id:playerId });
@@ -82,13 +64,6 @@ export const Host = () => {
           <div className={styles.button__horizontal}></div>
           <div className={styles.button__vertical}></div>
         </button>
-        <audio 
-          id="player"
-          className="player"
-          controls={false} 
-          src={currentTrack.preview_url}>
-            
-        </audio>
       
         <button
           className={styles.button}
@@ -111,3 +86,5 @@ export const Host = () => {
     </div>
   );
 };
+
+export default Host;
